Improvements to Neo4J OGM
This PR enables proper serialization of the `Scope` objects. To enable this, we moved the scopes from `.passes.scopes` to `.graph`, which is a better place for them anyway.
