+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/helpers/NameConverter.kt
@@ -37,7 +37,7 @@ import org.neo4j.ogm.typeconversion.CompositeAttributeConverter
  *
  * Additionally, it converts the aforementioned Neo4J attributes in a node back into a [Name].
  */
-class NameConverter : CompositeAttributeConverter<Name> {
+class NameConverter : CompositeAttributeConverter<Name?> {
 
     companion object {
         const val FIELD_FULL_NAME = "fullName"
@@ -46,21 +46,25 @@ class NameConverter : CompositeAttributeConverter<Name> {
         const val FIELD_NAME_DELIMITER = "nameDelimiter"
     }
 
-    override fun toGraphProperties(value: Name): MutableMap<String, *> {
+    override fun toGraphProperties(value: Name?): MutableMap<String, *> {
         val map = mutableMapOf<String, String>()
-        // The full name of the node
-        map[FIELD_FULL_NAME] = value.toString()
 
-        // The local name of the node
-        map[FIELD_LOCAL_NAME] = value.localName
+        if (value != null) {
+            // The full name of the node
+            map[FIELD_FULL_NAME] = value.toString()
 
-        // The delimiter
-        map[FIELD_NAME_DELIMITER] = value.delimiter
+            // The local name of the node
+            map[FIELD_LOCAL_NAME] = value.localName
 
-        // For reasons such as backwards compatibility and the fact that Neo4J likes to display
-        // nodes in the UI with a "name" field as default, we also persist the full name (aka the
-        // toString() representation) as "name"
-        map[FIELD_NAME] = value.toString()
+            // The delimiter
+            map[FIELD_NAME_DELIMITER] = value.delimiter
+
+            // For reasons such as backwards compatibility and the fact that Neo4J likes to display
+            // nodes in the UI with a "name" field as default, we also persist the full name (aka
+            // the
+            // toString() representation) as "name"
+            map[FIELD_NAME] = value.toString()
+        }
 
         return map
     }
