+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/scopes/GlobalScope.kt
@@ -23,18 +23,26 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.scopes
+package de.fraunhofer.aisec.cpg.graph.scopes
 
-import de.fraunhofer.aisec.cpg.graph.statements.BreakStatement
-import de.fraunhofer.aisec.cpg.graph.statements.CompoundStatement
-
-class BlockScope(blockStatement: CompoundStatement) :
-    ValueDeclarationScope(blockStatement), Breakable {
-    private val breaks: MutableList<BreakStatement> = ArrayList()
-    override fun addBreakStatement(breakStatement: BreakStatement) {
-        breaks.add(breakStatement)
+class GlobalScope : StructureDeclarationScope(null) {
+    /**
+     * This should ideally only be called once. It constructs a new global scope, which is not
+     * associated to any AST node. However, depending on the language, a language frontend can
+     * explicitly set the ast node using [ScopeManager.resetToGlobal] if the language needs a global
+     * scope that is restricted to a translation unit, i.e. C++ while still maintaining a unique
+     * list of global variables.
+     */
+    fun mergeFrom(others: Collection<GlobalScope>) {
+        for (other in others) {
+            structureDeclarations.addAll(other.structureDeclarations)
+            valueDeclarations.addAll(other.valueDeclarations)
+            typedefs.putAll(other.typedefs)
+            // TODO what to do with astNode?
+            for (child in other.children) {
+                child.parent = this
+                children.add(child)
+            }
+        }
     }
-
-    override val breakStatements: List<BreakStatement>
-        get() = breaks
 }
