+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/scopes/Scope.kt
@@ -23,31 +23,52 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.scopes
+package de.fraunhofer.aisec.cpg.graph.scopes
 
+import com.fasterxml.jackson.annotation.JsonBackReference
 import de.fraunhofer.aisec.cpg.graph.Name
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.statements.LabelStatement
+import de.fraunhofer.aisec.cpg.helpers.NameConverter
+import org.neo4j.ogm.annotation.GeneratedValue
+import org.neo4j.ogm.annotation.Id
+import org.neo4j.ogm.annotation.NodeEntity
+import org.neo4j.ogm.annotation.Relationship
+import org.neo4j.ogm.annotation.typeconversion.Convert
 
 /**
  * Represent semantic scopes in the language. Depending on the language scopes can have visibility
  * restriction and can act as namespaces to avoid name collisions.
  */
-abstract class Scope(open var astNode: Node?) {
+@NodeEntity
+abstract class Scope(
+    @Relationship(value = "SCOPE", direction = Relationship.Direction.INCOMING)
+    @JsonBackReference
+    open var astNode: Node?
+) {
+
+    /** Required field for object graph mapping. It contains the scope id. */
+    @Id @GeneratedValue var id: Long? = null
 
     /** FQN Name currently valid */
     var scopedName: String? = null
 
     /** The real new name */
-    var name: Name? = null
+    @Convert(NameConverter::class) var name: Name? = null
 
     /**
      * Scopes are nested and therefore have a parent child relationship, this two members will help
      * navigate through the scopes,e.g. when looking up variables.
      */
+    @Relationship(value = "PARENT", direction = Relationship.Direction.OUTGOING)
     var parent: Scope? = null
+
+    /** The list of child scopes. */
+    @Transient
+    @Relationship(value = "PARENT", direction = Relationship.Direction.INCOMING)
     var children = mutableListOf<Scope>()
-    var labelStatements = mutableMapOf<String, LabelStatement>()
+
+    @Transient var labelStatements = mutableMapOf<String, LabelStatement>()
 
     fun addLabelStatement(labelStatement: LabelStatement) {
         labelStatement.label?.let { labelStatements[it] = labelStatement }
