+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/scopes/SwitchScope.kt
@@ -23,17 +23,14 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.scopes
+package de.fraunhofer.aisec.cpg.graph.scopes
 
-import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.statements.BreakStatement
-import de.fraunhofer.aisec.cpg.graph.types.Type
+import de.fraunhofer.aisec.cpg.graph.statements.SwitchStatement
 
-class TryScope(astNode: Node?) : ValueDeclarationScope(astNode), Breakable {
-    val catchesOrRelays = mutableMapOf<Type, MutableList<Node>>()
-
+class SwitchScope(switchStatement: SwitchStatement) :
+    ValueDeclarationScope(switchStatement), Breakable {
     private val breaks = mutableListOf<BreakStatement>()
-
     override fun addBreakStatement(breakStatement: BreakStatement) {
         breaks.add(breakStatement)
     }
