+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/scopes/TemplateScope.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Fraunhofer AISEC. All rights reserved.
+ * Copyright (c) 2021, Fraunhofer AISEC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,12 +23,8 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.scopes
+package de.fraunhofer.aisec.cpg.graph.scopes
 
 import de.fraunhofer.aisec.cpg.graph.Node
-import de.fraunhofer.aisec.cpg.graph.declarations.FunctionDeclaration
-import de.fraunhofer.aisec.cpg.graph.types.Type
 
-class FunctionScope(astNode: FunctionDeclaration) : ValueDeclarationScope(astNode) {
-    var catchesOrRelays = mutableMapOf<Type, MutableList<Node>>()
-}
+class TemplateScope(node: Node) : StructureDeclarationScope(node) {}
