+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/scopes/TryScope.kt
@@ -23,26 +23,21 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.scopes
+package de.fraunhofer.aisec.cpg.graph.scopes
 
-class GlobalScope : StructureDeclarationScope(null) {
-    /**
-     * This should ideally only be called once. It constructs a new global scope, which is not
-     * associated to any AST node. However, depending on the language, a language frontend can
-     * explicitly set the ast node using [ScopeManager.resetToGlobal] if the language needs a global
-     * scope that is restricted to a translation unit, i.e. C++ while still maintaining a unique
-     * list of global variables.
-     */
-    fun mergeFrom(others: Collection<GlobalScope>) {
-        for (other in others) {
-            structureDeclarations.addAll(other.structureDeclarations)
-            valueDeclarations.addAll(other.valueDeclarations)
-            typedefs.putAll(other.typedefs)
-            // TODO what to do with astNode?
-            for (child in other.children) {
-                child.parent = this
-                children.add(child)
-            }
-        }
+import de.fraunhofer.aisec.cpg.graph.Node
+import de.fraunhofer.aisec.cpg.graph.statements.BreakStatement
+import de.fraunhofer.aisec.cpg.graph.types.Type
+
+class TryScope(astNode: Node?) : ValueDeclarationScope(astNode), Breakable {
+    @Transient val catchesOrRelays = mutableMapOf<Type, MutableList<Node>>()
+
+    private val breaks = mutableListOf<BreakStatement>()
+
+    override fun addBreakStatement(breakStatement: BreakStatement) {
+        breaks.add(breakStatement)
     }
+
+    override val breakStatements: List<BreakStatement>
+        get() = breaks
 }
